@Jarcho 2022-04-02
The win isn't as big in practice as `cargo_metadata` is still used in `clippy_lints`, but `regex` is completely removed.