@xFrednet 2022-04-02
Everything looks good to me, thank you for this change. I think everyone is going to enjoy the faster compile time for `clippy_dev` :tada: 

@bors r+